Can now test either cli, or web AND cli
It's useful to test the cli without needing to install rails or mongodb.
You should still be able to run all the tests from the project root as
before, but now running rake or rspec from the CLI root runs just those
tests, without extra dependencies.

Also tries to make Travis test multiple versions of Ruby. The server
needs a particular version, but the CLI is expected to run on many
versions. On the non-server versions, Travis should only run the CLI
tests.