Bump rails to version 3.2.22.5
this is an easy upgrade. not a whole lot changed from rails 3.1 to 3.2. 3.2 offered some speed improvements in dev mode and routing. there's a few active record changes that won't apply to us since we're using mongoid so I left out those configurations. there's a few deprecations that popped up from this bump so it looks like we'll need to bump the mongo family of gems as a next step.