Set eager_load configurations and fix class loading issue
we were including `VimGolf` (the gem) in the top level in the keylog spec which was causing `VimGolf::Challenge` from the gem to replace the `Challenge` rails model constant. this was causing failures in any rails spec that used the `Challenge` model. we had a few options for resolving this. i decided we didn't need to include the module at all (we could have done it inside the describe block), but instead fully qualify all classes from the gem with the proper namespace