Support neovim in vimgolf cli
The '--nofork' option is invalid for neovim. 
It is also stated in comment on line 106 that '--nofork' option is only required for gvim.
So, it is more logical to have this option for gvim only not for vim or neovim. 
This code is tested with 'vim', 'neovim' and 'gvim' (vim-gtk3) and it is working without any error.