Display challenges in user page in chronological order
Make it consistent, so that refreshing the page will not show them in a
different order. The ordering matches the one for the challenges at the
root page.

Add a spec test for player_best_scores with two challenges, which is now
possible since the resulting ordering is predictable. (The test is flaky
without the corresponding change to the Mongo query.)

Benchmarked this change on a page, timing went from (0.25547s ± 0.00663s)
before this change to (0.26706s ± 0.01071s) after it, which I would deem
acceptable.