Add query for entries per user for a challenge
The query ranks all entries by a player among the best entry of each
other user, for that challenge. That way we can both get the ranking of
the best entry for the user and we can also find what the ranking would
have been for each other entry, was that the best entry for that user.

Add a spec test case for the query.