Ignore codeclimate warnings for app/repositories/
This file mostly consists of MongoDB queries, encoded in pseudo-JSON format
with little Ruby logic in it. So there's little point in raising warnings for
code duplication (queries should be as explicit as possible), number of lines
in methods or number of lines in the file in general.

Rather than messing with the codeclimate global options (such as mass threshold
for code duplication), let's instead exclude this file from the code climate
checks.