Display all challenge solutions per user (#294)
* Refactor challenge snippets into separate views
* Introduce a highlight_owner? to submissions

So that we can selectively disable it for per-user challenge data.

* Add query for entries per user for a challenge

The query ranks all entries by a player among the best entry of each
other user, for that challenge. That way we can both get the ranking of
the best entry for the user and we can also find what the ranking would
have been for each other entry, was that the best entry for that user.

Add a spec test case for the query.

* Introduce a page with per-user entries for a challenge

The page is rendered at the /challenges/:id/user/:user path. The
rendered page is similar to a challenge page, but it only shows the
entries by a single player. It also shows *all* entries by that player.

You can see the score for all entries (where the Leaderboard usually
is), but same as on a challenge page, you can only see the actual script
submissions for entries if you have a score good enough to be able to
see them (or if you're the creator of the challenge, or if you're an
admin.)

I added a link from number of "Attempts:" on a user page to show the
entries submitted by that user for that particular challenge. That
looked like a natural place to link to it. (Perhaps linking from the
challenge page to the entries of a specific user would also be nice.)

The actual ranking position for entries is only really valid for the
best entry of a user, since the ranking is for *users* and not for
entries. But to approximate how good the other entries are, I calculated
what would be their ranking position if that was the best submission by
that user and I'm displaying those as the rankings for these
submissions. In order to mark them as not being the actual ranking, I'm
rendering those as #>n instead of #n (e.g. #>15 instead of #15, for an
entry that comes right after the user that is currently #15 in that
challenge.)

Caveat: There's no pagination in the per-user page. I thought it was
unlikely that a single user would have over 100 entries, so I ended up
skipping support for it.

* Link to the per-user entries for a challenge from best score

Add a link from the leaderboard, so it's easy to go from a challenge
page to the list of all entries by a single user.

* Ignore codeclimate warnings for app/repositories/

This file mostly consists of MongoDB queries, encoded in pseudo-JSON format
with little Ruby logic in it. So there's little point in raising warnings for
code duplication (queries should be as explicit as possible), number of lines
in methods or number of lines in the file in general.

Rather than messing with the codeclimate global options (such as mass threshold
for code duplication), let's instead exclude this file from the code climate
checks.