Update omniauth to 2.0, but keep allowing GET method
Codeclimate complains about omniauth 1.9.1, but my previous attempt to
upgrade it caused tests to fail. Looking into it, I noticed it's due to
a CSRF vulnerability designated CVE-2015-9284 which is addressed in
version 2.0 of omniauth.

Addressing that vulnerability requires using POST for authentication
requests, but it turns out that is hard to do, because either it
requires using `link_to ..., method: :post`, but that didn't work (at
least with RSpec tests, not sure why not, maybe JavaScript related), or
changing from `link_to` to `button_to`, but that affects the look of the
page and would require a non-trivial amount of CSS to style the buttons
the same as the links are today.

So I noticed I can still upgrade to 2.0 while keeping GET allowed by
tweaking the OmniAuth.config.allowed_request_methods variable for now,
and I decided to do that, to allow us to move forward, appease
Codeclimate and allow us to address the CSRF vulnerability (possibly by
switching to buttons for log-in) as a follow up.

Tested that RSpec tests still pass successfully after this commit.