Switch to using Challenge#top_entries instead of Mongo queries.
MongoDB queries were hitting sorting limits and disk use for sorting
isn't available in the MongoDB Atlas tier we're using.

So let's switch to just accessing Challenge#top_entries and do the
filtering in Ruby code. This is a little bit slower, but it should work
just fine for reasonably sized challenges. In my testing, I had trouble
with a 37,000 entry challenge (request exceeded 15,000ms) but up to
~30,000 entries it should work fine.