Switch Avatar to use an Identicon while Twitter profile pics are unavailable.
I found this service which provides a consistent identicon based on the
passed information. Pass it an MD5 digest of the username as the unique
information.

The Identicons are not as good as an actual profile picture... But
they're definitely better than serving a broken image (or removing the
avatars altogether.)