Downgrade mongoid + mongo + bson gems
The versions after the Ruby 2.6 upgrade started to choke on data from
specific entries in challenges, failing UTF-8 validation. The entries in
question have special characters such as '\x80' and '\xFF', so that's
not completely surprising.

The older versions of the gems work fine with that data, so let's
downgrade them in order to make the challenges accessible again.

Fixes #319.