Create an initial migration to create VimGolf tables
These will hold the Users, Challenges, Entries and Comments.

They mostly map the data stored in MongoDB collections.

The Challenges table also include a column for `legacy_urlkey` that
holds the MongoDB id of a specific challenge. Since that MongoDB id is
currently exposed as the URL of a challenge (or `vimgolf put` argument),
we will preserve it during the migration to keep URLs stable.

Comments will now store a reference to the Users table, rather than only
storing the textual nickname of the user who originally posted the
challenge (nicknames can actually change.)

Ideally some of the fields (such as Users.key and Users.nickname) should
have a unique index, but there are some data issues with repeated
entries for the same nickname or key, which we will address in a
follow-up after the migration from MongoDB.

Tested that this migration works on both sqlite3 and pg.

  $ bundle exec rake db:create db:migrate

The db/schema.rb file reflects a dump of the schema on sqlite3 (they
differ slightly), which is the backend primarily used in the development
(and test) environments.