Migrate the application from Mongoid to ActiveRecord
This commit updates the code to use ActiveRecord models, rather than
Mongoid documents, to implement the entities stored by the application.

For ranking entries (only the top entry by each user is presented), we
are using SQL window functions, which allow us to execute the bulk of
that filtering on the database side. Method Challenge#top_entries
implements query to the ranked entries for a challenge in a form that is
suitable for our purposes.

Window functions on PostgreSQL:
https://www.postgresql.org/docs/current/tutorial-window.html

Window functions on SQLite3:
https://sqlite.org/windowfunctions.html

NOTE: Window functions are only available in sqlite3 version 3.25+, so
developing and testing the app requires a system with a recent enough
version of sqlite3. For instance, Ubuntu 18.04 LTS ships sqlite3 3.22,
which doesn't have the necessary support. Ubuntu 20.04 LTS ships an
adequate sqlite3 version.

The MongoDB ids were exposed in the URLs for Challenges, so in order to
keep the URLs stable and keep backwards compatibility, we introduce a
new `urlkey` to encode the key to use for a Challenge URL.

For challenges migrated from MongoDB, column `legacy_urlkey` in the
database will hold that information.

For newly created challenges, we are using a 24-character similarly
looking key, using the `9v0` prefix, followed by 9 hex digits encoding
the `created_at` timestamp (so that URLs will sort in creation order)
followed by 13 hex digits encoding the ActiveRecord sequential id of the
Challenge. Prefix `9v0` is used because it's higher than any MongoDB ids
(which currently start with `4` and `5`), so it sorts higher, and the
`v` makes it look similar while not matching a hex digit. The trailing
`0` allows for some extension to change to a different scheme in the
future again, while still being able to preserve URLs and keep ordering.

Another difference from the MongoDB implementation is that the home page
will now display Challenges in strictly chornological order, newest to
oldest. The old code tried to prioritize "popular" challenges, but the
way I see it there were some shortcomings with that approach, for
instance a newly created Challenge would typically end up drifting to a
later page and not be visible enough. (We can address highlighting
"Popular" or "Featured" challenges in a follow-up.)

Finally, the MongoDB implementation would simply store the nickname of a
user in a Comment field. We're replacing that with a proper foreign key
to the User who initially posted the comment, in order to preserve
referential integrity.

This setup using ActiveRecord was tested locally, with a full import of
production MongoDB data from a couple of weeks ago, and proved to be
very responsive.

As an example, rendering the leaderboard for challenge "Simple,
Practical, and Common", with 32,662 entries, took around 600ms.

Rendering a user page for that same challenge, displaying all entries by
a single user, took under 300ms.

Here are the logs for these two accesses:

  I, [2021-04-29T11:29:16.460899 #1106821]  INFO -- : Processing by ChallengesController#show as HTML
  I, [2021-04-29T11:29:16.461014 #1106821]  INFO -- :   Parameters: {"id"=>"55b18bbea9c2c30d04000001"}
  I, [2021-04-29T11:29:17.022771 #1106821]  INFO -- : Completed 200 OK in 562ms (Views: 70.8ms | ActiveRecord: 486.7ms)
  127.0.0.1 - - [29/Apr/2021:11:29:17 -0700] "GET /challenges/55b18bbea9c2c30d04000001 HTTP/1.1" 200 - 0.5750

  I, [2021-04-29T11:30:59.762671 #1106821]  INFO -- : Processing by ChallengesController#user as HTML
  I, [2021-04-29T11:30:59.762731 #1106821]  INFO -- :   Parameters: {"id"=>"55b18bbea9c2c30d04000001", "username"=>"filbranden"}
  I, [2021-04-29T11:31:00.033980 #1106821]  INFO -- : Completed 200 OK in 271ms (Views: 29.7ms | ActiveRecord: 239.8ms)
  127.0.0.1 - - [29/Apr/2021:11:31:00 -0700] "GET /challenges/55b18bbea9c2c30d04000001/user/filbranden HTTP/1.1" 200 - 0.2776

It's quite possible there's more room for optimizations, but I'd say
this is already a terrific start!