Update tests to use the ActiveRecord models
Also update them to reflect the other changes in behavior introduced by
this chain of commits (such as Comment now referring to User, the new
`urlkey` for Challenges, etc.)

Since a significant amount of code was removed from RepositoryChallenge,
some test code was removed here. There were also 3 tests for the query
to return all entries for a user in a challenge that I temporarily
disabled since it will need to adapt.

Also removed all Rubocop exceptions for test code in `spec/` and
reformatted them appropriately in order to pass all the checks.

Updated the Rakefile to include creation of the `db/test.sqlite3` test
database (through execution of `db:test:prepare` target) upon running
tests, so that `bundle exec rake` will work out of the box on a freshly
checked out tree.

Also updated the Travis-CI configuration to use Ubuntu 20.04 LTS
(codename "focal") for test runs, since the SQL window functions require
a recent sqlite3 library version to work and the older Ubuntu versions
would fail tests for that reason (fail with SQLite3 exception on the
queries using window functions.)

Tested that the checks pass:

  $ bundle exec rake

Also confirmed that the tests work as expected on Travis-CI.