Rake task mongodb:import to migrate data from MongoDB
The task uses direct access to the MongoDB collections and creates
instances of ActiveRecord classes to import the data.

This task takes an optional `challenge_ids=...` argument, containing a
comma-separated list of MongoDB ids, in order to only migrate those
specific challenges. (Note that, in either case, the `mongodb:import`
task requires the database to be initially empty.) When importing only a
subset of all challenges, the task will also only import users on demand
as they're referred to by the challenge, entries or comments, so the
list of imported users will also be a partial one. This mode of
operation is useful to test real data on the Heroku Postgres free tier,
which has a limit for 10,000 rows, so wouldn't fit all entries (even all
entries of single very popular challenges), or all active golfers.

The `mongodb:import` task also handles connecting Comments back to Users
through a direct relationship, looking up the user by its nickname and
storing the ActiveRecord id for that user in the newly imported Comment.

Since nicknames can change (they're based on Twitter handles and users
are allowed to change those), some comments refer to nicknames that are
no longer valid users. In order to handle those during the migration,
we're creating "dummy" new users, using a provider `none` to tag these
cases. (We might need a follow-up with some data cleanup here, after the
migration out of MongoDB.)

The `mongodb:import` task also collects and reports statistics on the
data migration, to help inspection of progress and monitoring the
performance of the operation.

Tested migrations locally (to both SQLite3 and PostgreSQL) and on Heroku
with Postgres free tier (limited to 10,000 rows):

  1. Migrate 8 selected challenges into local SQLite3 database:

        I, [2021-04-28T15:48:45.681949 #1084283]  INFO -- : MongoImporter: Migration *COMPLETED*!

        --------------------------------------------------
        Users..............:    9098
        Challenges.........:       8 /       8 (  1%)
        Entries............:   65545
        Comments...........:     152

        Total Objects......:   74803
        Elapsed............:   5min 58sec
        --------------------------------------------------

   2. Migrate all challenges from a copy of the production MongoDB
      database as of a few weeks back, into a local SQLite3 database:

        I, [2021-04-28T16:23:46.085219 #1085304]  INFO -- : MongoImporter: Migration *COMPLETED*!

        --------------------------------------------------
        Users..............:   30687 /   30687 (  1%)
        Challenges.........:     513 /     513 (  1%)
        Entries............:  372776
        Comments...........:    3141

        Total Objects......:  407117
        Elapsed............:  32min 27sec
        --------------------------------------------------

   3. Migrate all challenges from a copy of the production MongoDB
      database as of a few weeks back, into a local Postgres database:

        I, [2021-04-28T18:05:53.863711 #1094634]  INFO -- : MongoImporter: Migration *COMPLETED*!

        --------------------------------------------------
        Users..............:   30687 /   30687 (100%)
        Challenges.........:     513 /     513 (100%)
        Entries............:  372776
        Comments...........:    3141

        Total Objects......:  407117
        Elapsed............:  14min 32sec
        --------------------------------------------------

  4. Migrate selected challenges, directly from the production MongoDB
     instance, into a Heroku Postgres free tier database, with the
     migration running on Heroku:

        $ heroku run --app vimgolf-staging bundle exec rails mongodb:import challenge_ids=5f1e0217becb80000692b9c4,5ebe8a63d8085e000c2f5bd5,6013804df3308e0009368f1c,51a388ac06f7d20002000006,592e0f555b60ba0009000001

        I, [2021-04-29T01:26:38.586319 #4]  INFO -- : MongoImporter: Migration *COMPLETED*!

        --------------------------------------------------
        Users..............:    1062
        Challenges.........:       5 /       5 (100%)
        Entries............:    4384
        Comments...........:      35

        Total Objects......:    5486
        Elapsed............:   1min 26sec
        --------------------------------------------------

The expectation is that the real data migration from MongoDB to Heroku
Postgres importing all data should take around 15 minutes (similar to
local Postgres) and shouldn't take much longer than 30 minutes.
Reporting will allow for monitoring progress of that migration, allowing
us to back off if it seems something is going unexpectedly.