Do not dump the schema on migration when using Postgres
The format of `db/schema.rb` is slightly different between SQLite and
Postgres and loading a schema dump from Postgres on SQLite produces a
non-working database.

As such, let's avoid dumping the schema when developing on Postgres and
keep the `db/schema.rb` file working with the SQLite format (which leads
to an easier ramp-up, almost zero setup required, for new developers.)
As a result, we're also avoiding using tasks such as `db:setup` and
`db:reset` on Postgres (since those rely on the `db/schema.rb`
contents.)

Tested:

  $ DATABASE_ADAPTER=pg bundle exec rails db:migrate:reset users=70 challenges=10 entries=80

Confirmed that `db/schema.rb` was unmodified.