Replace poltergeist+phantomjs with Cuprite
PhantomJS is no longer developed and Poltergeist is archived. The
recommended replacement seems to be Culprite, which is a headless Chrome
driver for Capybara.

(I ran into this issue after upgrading to Fedora 34 and phantomjs
stopped working due to the openssl version upgrade, so this was a good
opportunity to shed the dependency on a deprecated project.)

Tested:
$ bundle exec rspec