Include position in ranking for each challenge in user page
This reimplements PR #293 on top of ActiveRecord.

Also re-enable the test cases that were temporarily disabled. Two of
those were testing for this feature, the third one tested a related
feature (showing number of submissions for the same challenge) but it
was using a method removed during the ActiveRecord migration.

Tested:
- Accessing a user page with `?ranking=1` displays the position of the
  best user submission for each challenge.
- Tests under `bundle exec rspec` passing.