Add a test for `vimgolf put` using WebMock.
Use WebMock stubs to hook into the calls to download the challenge and
upload a submission. We also mock the homedir and inject a config.yaml
with a mock user apikey.

Tested that `bundle exec rspec` works correctly. Also tested this test
code with a commit that switches GOLFHOST to use https and confirmed
that the test keeps working as expected.

Use WebMock.allow_net_connect! after the CLI rspec is run, in order to
allow for net connections in other tests. (Some other tests make
connections to the real production vimgolf website, or other connections
to localhost, and need to be able to connect to the real backend.)