Fix ActiveRecord relationships on comments.
The has_many from User to Comment was missing. Furthermore, the comments
tests were creating test comments that did not belong to an user or an
entry. Enable belongs_to validation, which makes that test fail when not
including this fix.

Validation of belongs_to is the default in Rails 5.0 and updating
config/application.rb using `rails app:update` uncovered this issue.
Let's fix it in an initial commit, so that the `app:upgrade` commit
is cleaner.

Tested:
$ bundle exec rspec
$ bundle exec rubocop