Use Net::HTTP.start instead of .post and .get_response directly.
Older versions of Ruby don't have the new APIs, so let's stick with a
slightly more verbose API that still works on older versions of Ruby.
Confirmed that this works all the way back to Ruby 2.0.

We could drop support for older versions of Ruby, but since this is on
the client code only and it's not a huge effort to maintain it and keep
working on older Ruby, let's keep doing so for another short while...

CentOS 7 shipped Ruby 2.0 and Ubuntu 14.04 LTS shipped Ruby 1.9 by
default but also an option for Ruby 2.0, so it seems supporting all the
way to 2.0 *should* be enough for most cases...