Downgrade rexml for Ruby 2.0 in client test runs.
This only affects client *development* environment and is important for
tests. rexml is a dependency of webmock, that's how it's being pulled.
Newer versions of rexml (3.2.5 is latest, at the time of this writing)
are incompatible with Ruby 2.0 (but they don't seem to include a
requirement for the Ruby version that would have older versions pick a
Gem that would still work there.) Since we would like to have the tests
working on older Ruby, let's force a Gem version for this dependency
that will fix it there.