Use .strip_heredoc instead of squigglies, for compat with Ruby 2.0.
This only affects test cases. The syntax of using a squiggly with the
ending token of a here document was introduced in Ruby 2.3. For older
versions, there's `strip_heredoc`.

It seems Rails should provide those in activesupport, but I didn't
manage to get it to work... Still, defining that one myself wasn't too
much trouble and since this is pretty isolated to a single test case I
went ahead with this approach.