Use ruby_identicon to produce an identicon as a data URL
This removes an external dependency that has been broken for a while.

The ruby_identicon gem produces a PNG image in a format that is suitable to use
in a data: URL, which also reduces the number of HTTP connections needed to
properly render a page.

Tweak the identicon size to match our 48x48 space reserved for images.

Tested: Actually pushed to both vimgolf-staging and production already.