Add Dockerfile for container image to run the vimgolf client
Build an image to run the client on a container from the main repo.

Based on repository Hettomei/docker-vimgolf which was based on kramos/vimgolf.
Let's make it official, store it in GitHub Container Registry and add a
GitHub Action to rebuild it periodically.

Optimize it by using debian:stable-slim as the initial image. Also remove
apt caches after installing the required Ruby and Vim packages.

Tested:

```
% podman build -t local-vimgolf client/
$ podman run -it --rm -e key=... local-vimgolf 9v0062e3e96a000000000226
```