Drop dependencies on MongoDB libraries
We're now on Postgres, so drop the gems that are related to MongoDB.

This uncovered a few issues, mainly code that was working around some MongoDB
BSON encoding, also using the Symbol#in method for ActiveRecord where() queries
and also some simplification was possible by dropping Keylog#parse in favor
of using new() directly.

Tested: `bundle exec rspec` passes. Also pushed the code to vimgolf-staging
and confirmed it works as exepcted by navigating and using the website, also
playing some challenges.